feat: 공휴일 삭제 예외 처리

+++ src/main/java/toy/test/holidaymanager/holiday/adapter/out/persistence/HolidayAdapter.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import toy.test.holidaymanager.holiday.adapter.out.persistence.entity.HolidayJpaEntity;
+import toy.test.holidaymanager.holiday.adapter.out.persistence.exception.HolidayNotFoundException;
 import toy.test.holidaymanager.holiday.adapter.out.persistence.mapper.HolidayJpaMapper;
 import toy.test.holidaymanager.holiday.adapter.out.persistence.repository.HolidayJpaRepository;
 import toy.test.holidaymanager.holiday.application.port.in.command.RemoveCommand;
@@ -42,10 +43,14 @@ public class HolidayAdapter implements HolidayRepository {
     @Transactional
     @Override
     public void deleteAllByCondition(final RemoveCommand command) {
-        List<HolidayJpaEntity> entities = holidayJpaRepository.findAllByYearAndCountryCode(
-                command.year().value(),
-                command.countryCode().value()
-        );
+        int year = command.year().value();
+        String countryCode = command.countryCode().value();
+        List<HolidayJpaEntity> entities = holidayJpaRepository.findAllByYearAndCountryCode(year, countryCode);
+
+        if (entities.isEmpty()) {
+            throw new HolidayNotFoundException(year, countryCode);
+        }
+
         holidayJpaRepository.deleteAll(entities);
     }
 }
